Added large new example for GIS faceting based on Bureau of Labor Statistics, National Atlas, and US Census data.
To load the new dataset, 'cd data', then 'rake'.  Files are downloaded by ftp and processed for use in the faceting database.

This revision is intended to illustrate use of the new GIS faceting extensions to repertoire-faceting, using Google Earth.  Once installed, the GIS demo allows the user to explore a set employment statistics showing in which states and counties different businesses are found.